Move OlAsyncIOScheduler to shared utils dir
