+++ scripts/monitoring/tests/test_utils_py.py
@@ -1,6 +1,7 @@
 from unittest.mock import patch
 
-from scripts.monitoring.utils import OlAsyncIOScheduler, bash_run, limit_server
+from scripts.monitoring.utils import bash_run, limit_server
+from scripts.utils.scheduler import OlAsyncIOScheduler
 
 
 def test_bash_run():
@@ -25,7 +26,7 @@ def test_bash_run():
 
 def test_limit_server():
     with patch("os.environ.get", return_value="allowed-server"):
-        scheduler = OlAsyncIOScheduler()
+        scheduler = OlAsyncIOScheduler("X")
 
         @limit_server(["allowed-server"], scheduler)
         @scheduler.scheduled_job("interval", seconds=60)
@@ -36,7 +37,7 @@ def test_limit_server():
         assert scheduler.get_job("sample_job") is not None
 
     with patch("os.environ.get", return_value="other-server"):
-        scheduler = OlAsyncIOScheduler()
+        scheduler = OlAsyncIOScheduler("X")
 
         @limit_server(["allowed-server"], scheduler)
         @scheduler.scheduled_job("interval", seconds=60)
@@ -47,7 +48,7 @@ def test_limit_server():
         assert scheduler.get_job("sample_job") is None
 
     with patch("os.environ.get", return_value="allowed-server0"):
-        scheduler = OlAsyncIOScheduler()
+        scheduler = OlAsyncIOScheduler("X")
 
         @limit_server(["allowed-server*"], scheduler)
         @scheduler.scheduled_job("interval", seconds=60)
@@ -58,7 +59,7 @@ def test_limit_server():
         assert scheduler.get_job("sample_job") is not None
 
     with patch("os.environ.get", return_value="ol-web0.us.archive.org"):
-        scheduler = OlAsyncIOScheduler()
+        scheduler = OlAsyncIOScheduler("X")
 
         @limit_server(["ol-web0"], scheduler)
         @scheduler.scheduled_job("interval", seconds=60)
